Move `CamelCasedName` out of Java package
