+++ base/src/main/java/io/spine/code/java/MethodName.java
@@ -20,6 +20,7 @@
 
 package io.spine.code.java;
 
+import io.spine.code.CamelCasedName;
 import io.spine.value.StringTypeValue;
 
 import java.lang.reflect.Method;
